add to string for data models and move to service api package
