--- src/test/java/io/rsocket/transport/netty/GreetingResponse.java
@@ -1,11 +1,0 @@
-package io.rsocket.transport.netty;
-
-public class GreetingResponse {
-
-	private CharSequence greeting;
-
-	public CharSequence greeting() {
-		return greeting;
-	}
-
-}
