+++ src/test/java/io/rsocket/transport/netty/api/GreetingRequest.java
@@ -1,6 +1,11 @@
-package io.rsocket.transport.netty;
+package io.rsocket.transport.netty.api;
 
 public class GreetingRequest {
+
+	@Override
+	public String toString() {
+		return "GreetingRequest [name=" + name + "]";
+	}
 
 	private String name;
