+++ src/test/java/io/rsocket/transport/netty/api/GreetingResponse.java
@@ -1,0 +1,16 @@
+package io.rsocket.transport.netty.api;
+
+public class GreetingResponse {
+
+	@Override
+	public String toString() {
+		return "GreetingResponse [greeting=" + greeting + "]";
+	}
+
+	private CharSequence greeting;
+
+	public CharSequence greeting() {
+		return greeting;
+	}
+
+}
